move resource_helpers.go to a subpackage
