+++ pkg/api/resource/resource_helpers_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package api
+package resource
 
 import (
 	"testing"
@@ -22,13 +22,14 @@ import (
 
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/kubernetes/pkg/api"
 )
 
 func TestResourceHelpers(t *testing.T) {
 	cpuLimit := resource.MustParse("10")
 	memoryLimit := resource.MustParse("10G")
-	resourceSpec := ResourceRequirements{
-		Limits: ResourceList{
+	resourceSpec := api.ResourceRequirements{
+		Limits: api.ResourceList{
 			"cpu":             cpuLimit,
 			"memory":          memoryLimit,
 			"kube.io/storage": memoryLimit,
@@ -40,8 +41,8 @@ func TestResourceHelpers(t *testing.T) {
 	if res := resourceSpec.Limits.Memory(); res.Cmp(memoryLimit) != 0 {
 		t.Errorf("expected memorylimit %v, got %v", memoryLimit, res)
 	}
-	resourceSpec = ResourceRequirements{
-		Limits: ResourceList{
+	resourceSpec = api.ResourceRequirements{
+		Limits: api.ResourceList{
 			"memory":          memoryLimit,
 			"kube.io/storage": memoryLimit,
 		},
@@ -55,7 +56,7 @@ func TestResourceHelpers(t *testing.T) {
 }
 
 func TestDefaultResourceHelpers(t *testing.T) {
-	resourceList := ResourceList{}
+	resourceList := api.ResourceList{}
 	if resourceList.Cpu().Format != resource.DecimalSI {
 		t.Errorf("expected %v, actual %v", resource.DecimalSI, resourceList.Cpu().Format)
 	}
@@ -64,57 +65,20 @@ func TestDefaultResourceHelpers(t *testing.T) {
 	}
 }
 
-func newPod(now metav1.Time, ready bool, beforeSec int) *Pod {
-	conditionStatus := ConditionFalse
+func newPod(now metav1.Time, ready bool, beforeSec int) *api.Pod {
+	conditionStatus := api.ConditionFalse
 	if ready {
-		conditionStatus = ConditionTrue
+		conditionStatus = api.ConditionTrue
 	}
-	return &Pod{
-		Status: PodStatus{
-			Conditions: []PodCondition{
+	return &api.Pod{
+		Status: api.PodStatus{
+			Conditions: []api.PodCondition{
 				{
-					Type:               PodReady,
+					Type:               api.PodReady,
 					LastTransitionTime: metav1.NewTime(now.Time.Add(-1 * time.Duration(beforeSec) * time.Second)),
 					Status:             conditionStatus,
 				},
 			},
 		},
-	}
-}
-
-func TestIsPodAvailable(t *testing.T) {
-	now := metav1.Now()
-	tests := []struct {
-		pod             *Pod
-		minReadySeconds int32
-		expected        bool
-	}{
-		{
-			pod:             newPod(now, false, 0),
-			minReadySeconds: 0,
-			expected:        false,
-		},
-		{
-			pod:             newPod(now, true, 0),
-			minReadySeconds: 1,
-			expected:        false,
-		},
-		{
-			pod:             newPod(now, true, 0),
-			minReadySeconds: 0,
-			expected:        true,
-		},
-		{
-			pod:             newPod(now, true, 51),
-			minReadySeconds: 50,
-			expected:        true,
-		},
-	}
-
-	for i, test := range tests {
-		isAvailable := IsPodAvailable(test.pod, test.minReadySeconds, now)
-		if isAvailable != test.expected {
-			t.Errorf("[tc #%d] expected available pod: %t, got: %t", i, test.expected, isAvailable)
-		}
 	}
 }
